feat: complete backup scheduler redesign with legacy API removal
This commit finalizes the backup scheduler redesign implementation:

**Legacy API Cleanup:**
- Remove all legacy scheduler endpoints from backup router
- Delete old backup_scheduler.py service implementation
- Remove legacy scheduler tests
- Clean up unused imports and dependencies

**Service Unification:**
- Rename new_backup_scheduler.py to backup_scheduler.py
- Update BackupSchedulerService class name (remove "New" prefix)
- Consolidate singleton instance naming for consistency
- Add start_scheduler/stop_scheduler methods for main.py integration

**Documentation Updates:**
- Mark Phase 1-2 as completed in redesign specification
- Add operational guide with API usage examples
- Update implementation status to "ready for production"
- Document legacy API removal completion

**System Integration:**
- Ensure proper startup/shutdown lifecycle in main.py
- Update all import paths across the codebase
- Maintain backward compatibility for service interface
- Complete cache management and error handling

The new database-persistent backup scheduler is now fully operational
with improved permissions, server state awareness, and comprehensive
audit logging. All legacy memory-based scheduler code has been removed.
