+++ app/backups/router.py
@@ -1,4 +1,3 @@
-from typing import Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlalchemy.orm import Session
@@ -22,7 +21,7 @@ from app.core.exceptions import (
     FileOperationException,
     ServerNotFoundException,
 )
-from app.servers.models import BackupType, Server
+from app.servers.models import BackupType
 from app.services.authorization_service import authorization_service
 from app.services.backup_service import backup_service
 from app.users.models import Role, User
@@ -505,266 +504,3 @@ async def create_scheduled_backups(
         )
 
 
-# Backup Scheduler Management Endpoints
-
-
-@router.get("/scheduler-legacy/status")
-async def get_scheduler_status(current_user: User = Depends(get_current_user)):
-    """
-    Get backup scheduler status (admin only) - LEGACY ENDPOINT
-
-    Returns the current status of the backup scheduler including
-    scheduled servers and their next backup times.
-    """
-    try:
-        # Only admins can view scheduler status
-        if current_user.role != Role.admin:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail="Only admins can view scheduler status",
-            )
-
-        from app.services.backup_scheduler import backup_scheduler
-
-        scheduler_status = backup_scheduler.get_scheduler_status()
-        return scheduler_status
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Failed to get scheduler status: {str(e)}",
-        )
-
-
-@router.post("/scheduler-legacy/servers/{server_id}/schedule")
-async def add_server_to_schedule(
-    server_id: int,
-    interval_hours: int = Query(24, ge=1, le=168, description="Backup interval in hours"),
-    max_backups: int = Query(7, ge=1, le=30, description="Maximum backups to keep"),
-    current_user: User = Depends(get_current_user),
-    db: Session = Depends(get_db),
-):
-    """
-    Add server to backup schedule (admin only) - LEGACY ENDPOINT
-
-    Schedules automatic backups for the specified server.
-
-    - **interval_hours**: How often to create backups (1-168 hours)
-    - **max_backups**: Maximum number of backups to keep (1-30)
-    """
-    try:
-        # Only admins can manage scheduler
-        if current_user.role != Role.admin:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail="Only admins can manage backup schedules",
-            )
-
-        # Check if server exists
-        server = (
-            db.query(Server)
-            .filter(Server.id == server_id, Server.is_deleted.is_(False))
-            .first()
-        )
-        if not server:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND, detail="Server not found"
-            )
-
-        from app.services.backup_scheduler import backup_scheduler
-
-        backup_scheduler.add_server_schedule(
-            server_id=server_id,
-            interval_hours=interval_hours,
-            max_backups=max_backups,
-            enabled=True,
-        )
-
-        return {
-            "message": f"Server {server_id} added to backup schedule",
-            "interval_hours": interval_hours,
-            "max_backups": max_backups,
-        }
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Failed to add server to schedule: {str(e)}",
-        )
-
-
-@router.put("/scheduler-legacy/servers/{server_id}/schedule")
-async def update_server_schedule(
-    server_id: int,
-    interval_hours: Optional[int] = Query(
-        None, ge=1, le=168, description="Backup interval in hours"
-    ),
-    max_backups: Optional[int] = Query(
-        None, ge=1, le=30, description="Maximum backups to keep"
-    ),
-    enabled: Optional[bool] = Query(None, description="Enable/disable scheduled backups"),
-    current_user: User = Depends(get_current_user),
-    db: Session = Depends(get_db),
-):
-    """
-    Update server backup schedule (admin only) - LEGACY ENDPOINT
-
-    Updates the backup schedule settings for a server.
-    Creates a new schedule if none exists for the server.
-    """
-    try:
-        # Only admins can manage scheduler
-        if current_user.role != Role.admin:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail="Only admins can manage backup schedules",
-            )
-
-        # Check if server exists in database
-        server = (
-            db.query(Server)
-            .filter(Server.id == server_id, Server.is_deleted.is_(False))
-            .first()
-        )
-        if not server:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND, detail="Server not found"
-            )
-
-        from app.services.backup_scheduler import backup_scheduler
-
-        schedule = backup_scheduler.get_server_schedule(server_id)
-
-        # If no schedule exists, create a new one with default values
-        if not schedule:
-            # Set default values for new schedule
-            default_interval = interval_hours or 24
-            default_max_backups = max_backups or 7
-            default_enabled = enabled if enabled is not None else True
-
-            backup_scheduler.add_server_schedule(
-                server_id=server_id,
-                interval_hours=default_interval,
-                max_backups=default_max_backups,
-                enabled=default_enabled,
-            )
-
-            updated_schedule = backup_scheduler.get_server_schedule(server_id)
-            return {
-                "message": f"Created new backup schedule for server {server_id}",
-                "schedule": updated_schedule,
-            }
-
-        # Update existing schedule
-        backup_scheduler.update_server_schedule(
-            server_id=server_id,
-            interval_hours=interval_hours,
-            max_backups=max_backups,
-            enabled=enabled,
-        )
-
-        updated_schedule = backup_scheduler.get_server_schedule(server_id)
-        return {
-            "message": f"Updated backup schedule for server {server_id}",
-            "schedule": updated_schedule,
-        }
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Failed to update server schedule: {str(e)}",
-        )
-
-
-@router.get("/scheduler-legacy/servers/{server_id}/schedule")
-async def get_server_schedule(
-    server_id: int,
-    current_user: User = Depends(get_current_user),
-    db: Session = Depends(get_db),
-):
-    """
-    Get backup schedule for a specific server - LEGACY ENDPOINT
-
-    Returns the backup schedule configuration for the specified server.
-    Users can view schedules for servers they own, admins can view any schedule.
-    """
-    try:
-        # Check server access
-        authorization_service.check_server_access(server_id, current_user, db)
-
-        from app.services.backup_scheduler import backup_scheduler
-
-        schedule = backup_scheduler.get_server_schedule(server_id)
-
-        # Return 200 with null values if no schedule exists
-        if not schedule:
-            return {
-                "server_id": server_id,
-                "interval_hours": None,
-                "max_backups": None,
-                "enabled": False,
-                "last_backup": None,
-                "next_backup": None,
-            }
-
-        return {
-            "server_id": server_id,
-            "interval_hours": schedule["interval_hours"],
-            "max_backups": schedule["max_backups"],
-            "enabled": schedule["enabled"],
-            "last_backup": schedule.get("last_backup"),
-            "next_backup": schedule.get("next_backup"),
-        }
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Failed to get server schedule: {str(e)}",
-        )
-
-
-@router.delete("/scheduler-legacy/servers/{server_id}/schedule")
-async def remove_server_from_schedule(
-    server_id: int, current_user: User = Depends(get_current_user)
-):
-    """
-    Remove server from backup schedule (admin only) - LEGACY ENDPOINT
-
-    Removes a server from the automatic backup schedule.
-    """
-    try:
-        # Only admins can manage scheduler
-        if current_user.role != Role.admin:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail="Only admins can manage backup schedules",
-            )
-
-        from app.services.backup_scheduler import backup_scheduler
-
-        schedule = backup_scheduler.get_server_schedule(server_id)
-        if not schedule:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Server not found in backup schedule",
-            )
-
-        backup_scheduler.remove_server_schedule(server_id)
-
-        return {"message": f"Server {server_id} removed from backup schedule"}
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Failed to remove server from schedule: {str(e)}",
-        )
