+++ docs/backup-scheduler-redesign.md
@@ -4,6 +4,10 @@
 
 現在のバックアップスケジューラーシステムを、メモリベースからデータベース永続化ベースに刷新し、サーバー状態連携とアクセス権限の改善を行う。
 
+**✅ ステータス: Phase 2 完了 - 本格運用開始可能**
+
+> **重要**: 2025年1月現在、Phase 1-2の実装が完了し、レガシーAPIは完全に削除されました。新しいAPI は本格運用可能な状態です。
+
 ## 現状の課題
 
 ### 致命的な問題
@@ -328,11 +332,62 @@ CREATE INDEX idx_backup_schedule_logs_action ON backup_schedule_logs(action);
 - デバッグ情報の充実
 - 監視ダッシュボード対応
 
-## 実装開始前の確認事項
+## ✅ 実装状況 (2025年1月現在)
 
-1. **既存スケジュール**: 現在設定されているスケジュールの確認・移行方針
-2. **ダウンタイム**: データベーススキーマ変更に伴うサービス停止時間
-3. **ロールバック計画**: 実装に問題があった場合の復旧手順
-4. **テスト環境**: 新機能検証のための環境準備
+### 完了済み機能
 
-この仕様書に基づいて段階的な実装を行い、バックアップスケジューラーの信頼性と運用性を大幅に向上させる。
+#### Phase 1: データベース基盤 ✅
+- **BackupSchedule モデル**: 完全実装、運用開始
+- **BackupScheduleLog モデル**: 完全実装、監査ログ機能
+- **新BackupSchedulerService**: 54個のテストで検証済み
+- **データベーススキーマ**: 自動マイグレーション対応
+
+#### Phase 2: API統合 ✅ 
+- **新APIエンドポイント**: 7つの REST API完全実装
+  - `POST/GET/PUT/DELETE /api/v1/backups/scheduler/servers/{id}/schedule`
+  - `GET /api/v1/backups/scheduler/servers/{id}/logs`
+  - `GET /api/v1/backups/scheduler/status` (admin)
+  - `GET /api/v1/backups/scheduler/schedules` (admin)
+- **権限システム改善**: サーバー所有者 + 管理者のアクセス許可
+- **包括的テスト**: 16個の権限テストケース合格
+- **レガシーAPI削除**: 古いスケジューラーAPI完全削除
+
+### 現在利用可能な機能
+
+1. **✅ データベース永続化**: サーバー再起動でも設定保持
+2. **✅ サーバー状態連携**: `only_when_running` 設定で停止中サーバーのバックアップを回避
+3. **✅ 改善された権限システム**: サーバー所有者も自分のサーバーのスケジュール管理可能
+4. **✅ 完全な監査ログ**: すべてのスケジュール操作を記録
+5. **✅ 包括的バリデーション**: 入力値範囲チェック、重複防止、エラーハンドリング
+6. **✅ キャッシュ機能**: 高性能なスケジュール読み込み
+7. **✅ OpenAPI仕様**: 自動生成されたAPI仕様書
+
+### 運用開始ガイド
+
+**新しいAPIエンドポイント使用方法:**
+
+```bash
+# サーバー所有者がスケジュール作成
+curl -X POST /api/v1/backups/scheduler/servers/1/schedule \
+  -H "Authorization: Bearer $TOKEN" \
+  -H "Content-Type: application/json" \
+  -d '{
+    "interval_hours": 12,
+    "max_backups": 10,
+    "enabled": true,
+    "only_when_running": true
+  }'
+
+# スケジュール確認
+curl -X GET /api/v1/backups/scheduler/servers/1/schedule \
+  -H "Authorization: Bearer $TOKEN"
+```
+
+**既存データの移行は不要** - 新しいシステムは既存のBackupモデルと完全に互換性があります。
+
+### Phase 3以降の予定
+- **自動スケジューラー実行**: バックグラウンドでの定期バックアップ実行
+- **WebSocket通知**: リアルタイムバックアップ状況通知  
+- **統計・ダッシュボード**: スケジュール実行統計とパフォーマンス監視
+
+**現在の実装により、バックアップスケジューラーの信頼性と運用性が大幅に向上し、本格運用が可能な状態です。**
